chore: Implement SearchDialog component
Currently, this component is used by QuickSearch dialog. And then, it will be
used by the "Add to playlist" dialog.

I added some setters and getters to the QuickSearch dialog to abstract the interaction
between QuickSearch and SearchDialog from the Controller.

After this PR, I will implement the 'Add to playlist' dialog to incorporate the SearchDialog.